Revert "Topology Solver: Adjacency Graphs and Constraints API (#34385)"
This reverts commit 73054c1abb3f1592a27c6bba444146e1f2f57442.
