+++ tt_metal/fabric/CMakeLists.txt
@@ -76,7 +76,6 @@ target_sources(
         physical_system_descriptor.cpp
         topology_mapper.cpp
         topology_mapper_utils.cpp
-        topology_solver.cpp
 )
 
 # Include helper functions and generate headers from protobuf schemas
